Add todo model

+++ Models/Todo.php
@@ -5,10 +5,24 @@ class Todo
 {
     public function __construct(
         public ?int $id = null, 
-        public ?bool $isCompleted = null, 
-        public ?string $text = null)
+        public bool $isCompleted = false, 
+        public string $text = '')
     {
 
+    }
+
+    public static function all(PDO $pdo)
+    {
+        $query = "SELECT * FROM todos";
+        $stmt = $pdo->prepare($query);
+        $stmt->execute();
+        $result = $stmt->fetchAll();
+        $todos = [];
+        foreach ($result as $todo) 
+        {
+            $todos[] = new static($todo['id'], $todo['completed'], $todo['text']);
+        }
+        return $todos;
     }
 
     public static function find(PDO $pdo, int $todoId)
@@ -55,6 +69,10 @@ class Todo
 
     public function delete(PDO $pdo)
     {
+        if ($this->id === null) {
+            return null;
+        }
+
         $query = "DELETE FROM todos WHERE id = :todoId";
         $stmt = $pdo->prepare($query);
         $stmt->execute([
